had to redo showing image in cell because my work somehow didnt get commited last night
